refactor: convert file names to camel case
