+++ component/posting/hashTag.tsx
