+++ pages/join.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import { useState, useEffect } from 'react';
 import Link from 'next/link';
 import axios from 'axios';
-import LoginHeader from '../component/layout/login-header';
+import LoginHeader from '../component/layout/loginHeader';
 
 import { useRouter } from 'next/router';
